pyproject.toml
[project]
name = "log-ingestion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.135.3",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "uvicorn>=0.44.0",
]
